+++ test_playground/cli_init_demo/lib/app/config/network/interceptors/auth_interceptor.dart
@@ -1,0 +1,18 @@
+import 'package:dio/dio.dart';
+
+typedef TokenProvider = Future<String?> Function();
+
+class AuthInterceptor extends Interceptor {
+  final TokenProvider tokenProvider;
+
+  AuthInterceptor({required this.tokenProvider});
+
+  @override
+  void onRequest(RequestOptions options, RequestInterceptorHandler handler) async {
+    final token = await tokenProvider();
+    if (token != null && token.isNotEmpty) {
+      options.headers['Authorization'] = 'Bearer ' + token;
+    }
+    handler.next(options);
+  }
+}