+++ test_playground/cli_init_demo/lib/app/core/constants/app_constants.dart
@@ -1,0 +1,4 @@
+class AppConstants {
+  static const String appName = 'Flutter GetX App';
+  static const String version = '1.0.0';
+}