+++ test_playground/cli_init_demo/lib/app/core/constants/string_constants.dart
@@ -1,0 +1,5 @@
+class StringConstants {
+  static const String ok = 'OK';
+  static const String cancel = 'Cancel';
+  static const String confirm = 'Confirm';
+}