+++ test_playground/cli_init_demo/lib/app/routes/app_routes.dart
@@ -1,0 +1,5 @@
+part of 'app_pages.dart';
+
+abstract class AppRoutes {
+  AppRoutes._();
+}